feat: JWT Configuration 수정
