+++ src/main/java/planshare/server/planshare/user/service/UserService.java
@@ -10,14 +10,17 @@ import org.springframework.stereotype.Service;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestTemplate;
+import planshare.server.planshare.repository.MemberRepository;
+import planshare.server.planshare.security.JWTUtil;
 import planshare.server.planshare.user.dto.KakaoUserInfo;
 import planshare.server.planshare.user.dto.TokenResponse;
-import planshare.server.planshare.util.JWTUtil;
+
 
 @RequiredArgsConstructor
 @Service
 public class UserService {
 
+    private final MemberRepository memberRepository;
     private final JWTUtil jwtUtil;
 
     @Value("${client.id}")
